requirements.txt
psycopg2-binary>=2.9.3
click>=8.1.3
mimesis>=9.1.0
tabulate>=0.8.10
python-dotenv>=0.21.0
seaborn
